requirements.txt
openai
phidata>=0.0.55
pathlib
ipython
